Organized atoms and added documentation
