+++ packages/client/src/atoms/README.md
@@ -1,0 +1,12 @@
+### Atoms
+
+Atoms are the smallest, most basic building blocks of UI design. Typical examples of atoms we have in this project are:
+- Buttons
+- Labels
+- Links
+- Lists
+- Text renderers (Alert, ContactSentence, Disclaimer, TextToEdit)
+
+Usually, atoms are very abstract in their essence and limited to a single functionality that can be included in many pages. They are not very useful by themselves. 
+
+When adding new atoms, try to keep them as small and basic as possible. Being the smallest building block of atomic design, atoms cannot include other components.
