--- packages/client/src/components/FormStyles.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const StyledForm = styled.form`
-  width: 100%;
-  /* IE11 Fix */
-  flex-shrink: 0;
-`;
