+++ packages/client/src/components/Question/Question.tsx
@@ -3,14 +3,13 @@ import React, { FunctionComponent, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 
-import { ComponentWrapper } from "../../atoms";
+import { ComponentWrapper, Form } from "../../atoms";
 import { actions, eventNames } from "../../config/matomo";
 import { useTopic, useTopicData, useTracking } from "../../hooks";
 import { Answer } from "../../types";
 import { getAnswerLabel } from "../../utils";
 import { QUESTION_FORM } from "../../utils/test-ids";
 import Answers from "../Answers";
-import Form from "../Form";
 import Markdown from "../Markdown";
 import Modal from "../Modal";
 import Nav from "../Nav";
