+++ packages/client/src/pages/olo/OloLocationInput.tsx
@@ -3,8 +3,8 @@ import { Helmet } from "react-helmet";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
 
+import { Loading } from "../../atoms";
 import { TopicLayout } from "../../components/Layouts";
-import Loading from "../../components/Loading";
 import { LocationInput } from "../../components/Location/";
 import { useTopic } from "../../hooks";
 import { geturl, routes } from "../../routes";
